requirements.txt
aiohttp==3.5.4
aiomysql==0.0.20
asn1crypto==0.24.0
astroid==2.2.2
async-timeout==3.0.1
attrs==18.2.0
awesome-slugify==1.6.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
cryptography==2.4.2
Django==2.1.7
django-braces==1.13.0
idna==2.8
isort==4.3.10
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
multidict==4.5.2
mysql-connector==2.1.6
mysql-connector-python==8.0.13
mysqlclient==1.4.2.post1
Pillow==5.4.1
psutil==5.4.8
pycparser==2.19
pylint==2.3.1
PyMySQL==0.9.2
pytz==2018.9
redis==3.2.1
regex==2019.3.12
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.5.0
typed-ast==1.3.1
Unidecode==0.4.21
urllib3==1.24.1
virtualenv==16.2.0
wrapt==1.11.1
yarl==1.3.0
